market-update-function/requirements.txt
azure-common==1.1.28
azure-communication-email==1.0.0
azure-core==1.33.0
azure-mgmt-core==1.5.0
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
isodate==0.7.2
msrest==0.7.1
oauthlib==3.2.2
requests==2.32.3
requests-oauthlib==2.0.0
six==1.17.0
typing_extensions==4.13.1
urllib3==2.3.0
